requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
anthropic==0.34.2
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
